% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ComplaintsData}
\alias{ComplaintsData}
\title{\code{R6Class} for getting complaints data in a \code{ts} object}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
ComplaintsData
}
\description{
\code{R6Class} for getting complaints data in a \code{ts} object
}
\examples{
complaints.data <- ComplaintsData$new("~/workspaces/datakind-ws/cocUptoJuly2016.csv")
}
\keyword{datasets}

