% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoms.R
\name{detect_anomalies}
\alias{detect_anomalies}
\title{Detect anomalies in a given series}
\usage{
detect_anomalies(series, start_date = NULL, end_date = NULL)
}
\arguments{
\item{series, }{a time-series}

\item{start_date}{(optional) the start date to consider when detecting anoms}

\item{end_date}{(optional) the end date to consider when detecting anoms}
}
\description{
Detect anomalies in a given series
}

