% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/anoms.R
\name{detect_anomalies_1_day}
\alias{detect_anomalies_1_day}
\title{Detecting anomaly for a particular day.}
\usage{
detect_anomalies_1_day(subset_data, data_level, input_date, ...)
}
\arguments{
\item{series}{Hourly data}
}
\description{
Detecting anomaly for a particular day.
}

