% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/egovs_forecast.R
\name{egovs_forecasts}
\alias{egovs_forecasts}
\title{Extract forecasts for a given time series.}
\usage{
egovs_forecasts(series, ts_model, forecast_points = 3, as.df = TRUE,
  cleaned = FALSE, stl.decompose = FALSE, conf.intervals = c(80, 95),
  clip_negative = TRUE, ...)
}
\arguments{
\item{series}{a \code{ts} object}

\item{ts_model}{the time series model to use. ex. \code{arima}}

\item{forecast_points}{the number of periods to forecast}

\item{as.df}{whether to return the results as a df. If false, a \code{forecast} object is returned}

\item{cleaned}{if set to \code{TRUE}, the series is cleaned before forecasting}

\item{stl.decompose}{if set to \code{TRUE} the series is decomposed using STL}

\item{conf.intervals}{the confidence intervals. a vector of numbers between 1-100}

\item{clip_negative}{if set to \code{TRUE} negative values in forecasts are clipped to 0}

\item{...}{other model-specific params. see examples}
}
\value{
the forecasts
}
\description{
Extract forecasts for a given time series.
}
\examples{
egovs_forecasts(series, ts_model = "ets", ets.model="AAA", ets.damped=TRUE)
egovs_forecasts(series, ts_model = "arima", forecast_points = 5, arima.p = 3, arima.q = 2, arima.d = 1, lambda=NULL)
}

