% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{executeForecasts}
\alias{executeForecasts}
\title{Execute the full loop for forecasts}
\usage{
executeForecasts(configFilePath)
}
\arguments{
\item{configFilePath}{the path of the config file}
}
\description{
Execute the full loop for forecasts
}

