% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{export.series}
\alias{export.series}
\title{converts the series into a format that we can export}
\usage{
export.series(series, split.times = T)
}
\description{
converts the series into a format that we can export
}

