% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{forecasts.execute}
\alias{forecasts.execute}
\title{Execute the full loop for forecasts}
\usage{
forecasts.execute(configFilePath)
}
\arguments{
\item{configFilePath}{the path of the config file}
}
\description{
Execute the full loop for forecasts
}

