% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exec.R
\name{executeForecasts}
\alias{executeForecasts}
\title{Execute the full loop for forecasts}
\usage{
executeForecasts(configFilePath, outputFilePath)
}
\arguments{
\item{configFilePath}{the path of the config file}

\item{outputFilePath}{the path to write into. Will be overwritten if it already exists
converts the series into a format that we can export}
}
\description{
Execute the full loop for forecasts
}

